+++ applications/fishing-map/src/features/datasets/datasets.mock.ts
@@ -1,4 +1,5 @@
 import { Dataset } from '@globalfishingwatch/api-types/dist'
 
 export const datasets: Dataset[] = []
+
 export default datasets