+++ applications/fishing-map/src/features/map/popups/VesselEventsLayers.tsx
@@ -1,0 +1,61 @@
+import React, { Fragment } from 'react'
+import { useTranslation } from 'react-i18next'
+import { groupBy } from 'lodash'
+import { DateTime } from 'luxon'
+import { TooltipEventFeature } from 'features/map/map.hooks'
+import { formatI18nDate } from 'features/i18n/i18nDate'
+import styles from './Popup.module.css'
+
+// t('event.fishing', 'Fished for')
+// t('event.port', 'Anchored for')
+
+type ContextTooltipRowProps = {
+  features: TooltipEventFeature[]
+}
+
+function EnvironmentTooltipSection({ features }: ContextTooltipRowProps) {
+  const { t } = useTranslation()
+  const featuresByType = groupBy(features, 'layerId')
+  return (
+    <Fragment>
+      {Object.values(featuresByType).map((featureByType, index) => (
+        <div key={`${featureByType[0].title}-${index}`} className={styles.popupSection}>
+          <span
+            className={styles.popupSectionColor}
+            style={{ backgroundColor: featureByType[0].color }}
+          />
+          <div className={styles.popupSectionContent}>
+            {featureByType.map((feature, index) => {
+              const duration = DateTime.fromISO(feature.properties.end)
+                .diff(DateTime.fromISO(feature.properties.start), ['hours', 'minutes', 'seconds'])
+                .toObject()
+              return (
+                <Fragment>
+                  <div className={styles.row}>
+                    <span className={styles.rowText}>
+                      {formatI18nDate(feature.properties.start, { format: DateTime.DATETIME_FULL })}{' '}
+                    </span>
+                  </div>
+                  <div className={styles.row}>
+                    <span className={styles.rowText}>
+                      {t(`event.${feature.properties.type}Action` as any)}{' '}
+                      {duration.hours !== undefined &&
+                        duration.hours > 0 &&
+                        `${duration.hours} ${t('common.hour', {
+                          count: duration.hours,
+                        })} `}
+                      {duration.minutes &&
+                        `${duration.minutes} ${t('common.minute', { count: duration.minutes })}`}
+                    </span>
+                  </div>
+                </Fragment>
+              )
+            })}
+          </div>
+        </div>
+      ))}
+    </Fragment>
+  )
+}
+
+export default EnvironmentTooltipSection