+++ applications/fishing-map/src/features/timebar/timebar.selectors.ts
@@ -1,5 +1,12 @@
 import { createSelector } from '@reduxjs/toolkit'
-import { DatasetTypes, Resource, TrackResourceData } from '@globalfishingwatch/api-types'
+import { DateTime } from 'luxon'
+import {
+  ApiEvent,
+  DatasetTypes,
+  Resource,
+  ResourceStatus,
+  TrackResourceData,
+} from '@globalfishingwatch/api-types'
 import { resolveDataviewDatasetResource } from '@globalfishingwatch/dataviews-client'
 import { geoJSONToSegments, Segment } from '@globalfishingwatch/data-transforms'
 import { selectTimebarGraph } from 'features/app/app.selectors'
@@ -8,6 +15,7 @@ import {
   selectActiveVesselsDataviews,
 } from 'features/dataviews/dataviews.selectors'
 import { selectResources } from 'features/resources/resources.slice'
+import { EVENTS_COLORS } from 'data/config'
 
 type TimebarTrackSegment = {
   start: number
@@ -83,5 +91,91 @@ export const selectTracksGraphs = createSelector(
       }
     })
     return graphs
+  }
+)
+
+const selectEventsForTracks = createSelector(
+  [selectActiveTrackDataviews, selectResources],
+  (trackDataviews, resources) => {
+    const vesselsEvents = trackDataviews.map((dataview) => {
+      const { url: tracksUrl } = resolveDataviewDatasetResource(dataview, DatasetTypes.Tracks)
+      const { url: eventsUrl } = resolveDataviewDatasetResource(dataview, DatasetTypes.Events)
+      const hasEventData = eventsUrl && resources[eventsUrl]?.data
+      const tracksResourceResolved =
+        tracksUrl && resources[tracksUrl]?.status === ResourceStatus.Finished
+
+      // Waiting for the tracks resource to be resolved to show the events
+      if (!hasEventData || !tracksResourceResolved) {
+        return []
+      }
+      return (eventsUrl ? resources[eventsUrl].data : []) as ApiEvent[]
+    })
+    return vesselsEvents.filter((events) => events.length > 0)
+  }
+)
+
+interface RenderedEvent extends ApiEvent {
+  color: string
+  description: string
+}
+
+export const selectEventsWithRenderingInfo = createSelector(
+  [selectEventsForTracks],
+  (eventsForTrack) => {
+    const eventsWithRenderingInfo: RenderedEvent[][] = eventsForTrack.map(
+      (trackEvents: ApiEvent[]) => {
+        return (trackEvents || []).map((event: ApiEvent) => {
+          const vesselName = event.vessel.name || event.vessel.id
+
+          // TODO translate this
+          let description
+          switch (event.type) {
+            // case 'encounter':
+            //   if (event.encounter && event.encounter.vessel.name) {
+            //     description = `${vesselName} had an encounter with ${event.encounter.vessel.name}`
+            //   } else {
+            //     description = `${vesselName} had an encounter with another vessel`
+            //   }
+            //   break
+            // case 'port':
+            //   if (event.port && event.port.name) {
+            //     description = `${vesselName} docked at ${event.port.name}`
+            //   } else {
+            //     description = `${vesselName} Docked`
+            //   }
+            //   break
+            // case 'loitering':
+            //   description = `${vesselName} loitered`
+            //   break
+            case 'fishing':
+              description = `${vesselName} fishing`
+              break
+            default:
+              description = 'Unknown event'
+          }
+          const duration = DateTime.fromMillis(event.end as number)
+            .diff(DateTime.fromMillis(event.start as number), ['hours', 'minutes'])
+            .toObject()
+          description = `${description} for ${duration.hours}hrs ${Math.round(
+            duration.minutes as number
+          )}mns`
+
+          let colorKey = event.type as string
+          if (event.type === 'encounter') {
+            colorKey = `${colorKey}${event.encounter?.authorizationStatus}`
+          }
+          const color = EVENTS_COLORS[colorKey]
+          const colorLabels = EVENTS_COLORS[`${colorKey}Labels`]
+
+          return {
+            ...event,
+            color,
+            colorLabels,
+            description,
+          }
+        })
+      }
+    )
+    return eventsWithRenderingInfo
   }
 )