+++ applications/fishing-map/src/features/timebar/timebar.slice.ts
@@ -1,4 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
+import { ApiEvent } from '@globalfishingwatch/api-types/dist'
 import { RootState } from 'store'
 
 export type Range = {
@@ -8,11 +9,13 @@ export type Range = {
 
 type TimebarSlice = {
   highlightedTime: Range | undefined
+  highlightedEvent: ApiEvent | undefined
   hasChangedSettingsOnce: boolean
 }
 
 const initialState: TimebarSlice = {
   highlightedTime: undefined,
+  highlightedEvent: undefined,
   hasChangedSettingsOnce: false,
 }
 
@@ -23,6 +26,9 @@ const slice = createSlice({
     setHighlightedTime: (state, action: PayloadAction<Range>) => {
       state.highlightedTime = action.payload
     },
+    setHighlightedEvent: (state, action: PayloadAction<ApiEvent | undefined>) => {
+      state.highlightedEvent = action.payload
+    },
     disableHighlightedTime: (state) => {
       state.highlightedTime = undefined
     },
@@ -32,9 +38,12 @@ const slice = createSlice({
   },
 })
 
-export const { setHighlightedTime, disableHighlightedTime, changeSettings } = slice.actions
+export const { setHighlightedTime, setHighlightedEvent, disableHighlightedTime, changeSettings } =
+  slice.actions
+
 export default slice.reducer
 
 export const selectHighlightedTime = (state: RootState) => state.timebar.highlightedTime
+export const selectHighlightedEvent = (state: RootState) => state.timebar.highlightedEvent
 export const selectHasChangedSettingsOnce = (state: RootState) =>
   state.timebar.hasChangedSettingsOnce