+++ applications/dataviews-editor/CHANGELOG.md
@@ -1,9 +1,39 @@
 # @globalfishingwatchapp/dataviews-editor
 
+## 1.1.62
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @globalfishingwatch/dataviews-client@11.1.0
+  - @globalfishingwatch/layer-composer@8.1.0
+  - @globalfishingwatch/timebar@1.8.0
+
+## 1.1.61
+
+### Patch Changes
+
+- @globalfishingwatch/api-client@2.0.0
+- @globalfishingwatch/dataviews-client@11.0.0
+- @globalfishingwatch/layer-composer@8.0.0
+
+## 1.1.60
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+- Updated dependencies [undefined]
+  - @globalfishingwatch/layer-composer@7.3.0
+  - @globalfishingwatch/dataviews-client@10.0.0
+  - @globalfishingwatch/timebar@1.7.0
+
 ## 1.1.59
 
 ### Patch Changes
 
+- Updated dependencies [undefined]
+  - @globalfishingwatch/dataviews-client@9.0.0
+  - @globalfishingwatch/layer-composer@7.2.0
 - Updated dependencies [05fe3de6]
   - @globalfishingwatch/layer-composer@7.2.0
   - @globalfishingwatch/dataviews-client@9.0.0