+++ applications/sandbox/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 0.6.8
+
+### Patch Changes
+
+- @globalfishingwatch/api-client@2.0.0
+
 ## 0.6.7
 
 ### Patch Changes